Fixed issue in ending walk

+++ app/src/main/java/com/example/walklist/controllers/WalkController.kt
@@ -92,7 +92,7 @@ object WalkController: BaseController() {
         activeWalk.endPointLat = currentPointLat
         activeWalk.endPointLong = currentPointLong
         activeWalk.endAt = Date()
-        activeWalk.duration += (activeWalk.resumedAt.time - Date().time) / 1000 % 60
+        activeWalk.duration += ((activeWalk.resumedAt.time - Date().time) / 1000 % 60).toInt()
         activeWalk.distance += MapUtils.distanceBetween(activeWalk.resumedLat, activeWalk.resumedLong, activeWalk.endPointLat!!, activeWalk.endPointLong!!).toInt()
         activeWalk.encodedRoute = PolyUtils.append(activeWalk.encodedRoute, listOf(LatLng(activeWalk.endPointLat!!, activeWalk.endPointLong!!)))
